feat: reuse error map in astro db
